+++ lib/sourcescrub.rb
@@ -5,7 +5,9 @@ require 'sourcescrub/version'
 require 'sourcescrub/account'
 require 'sourcescrub/client'
 require 'sourcescrub/models'
+require 'sourcescrub/utils/veriables'
 
+# Sourcescrub
 module Sourcescrub
   TOKEN_URL = 'https://identity.sourcescrub.com'
   TOKEN_URI = '/connect/token'
@@ -14,5 +16,6 @@ module Sourcescrub
   API_URI   = 'https://api.sourcescrub.com/'
 
   class Error < StandardError; end
-  # Your code goes here...
+
+  include Utils::Veriables
 end