+++ lib/sourcescrub/models.rb
@@ -3,9 +3,11 @@
 module Sourcescrub
   # Models
   module Models
-    autoload :Entity,   'sourcescrub/models/concerns/entity'
-    autoload :Company,  'sourcescrub/models/company'
-    autoload :Source,   'sourcescrub/models/source'
-    autoload :Tag,      'sourcescrub/models/tag'
+    autoload :Entity,           'sourcescrub/models/concerns/entity'
+    autoload :Company,          'sourcescrub/models/company'
+    autoload :CompanyItems,     'sourcescrub/models/concerns/company_items'
+    autoload :Source,           'sourcescrub/models/source'
+    autoload :Tag,              'sourcescrub/models/tag'
+    autoload :Person,           'sourcescrub/models/person'
   end
 end