+++ lib/sourcescrub/models/employee.rb
@@ -1,0 +1,18 @@
+# frozen_string_literal: true
+
+module Sourcescrub
+  # Models
+  module Models
+    # Employee
+    class Employee < Entity
+      ENDPOINT = 'employees'
+
+      def field_ids
+        %w[
+          count
+          date
+        ]
+      end
+    end
+  end
+end