+++ lib/sourcescrub/models/financial.rb
@@ -1,0 +1,20 @@
+# frozen_string_literal: true
+
+module Sourcescrub
+  # Models
+  module Models
+    # Financial
+    class Financial < Entity
+      ENDPOINT = 'financials'
+
+      def field_ids
+        %w[
+          year
+          revenue
+          growth
+          growthTimePeriod
+        ]
+      end
+    end
+  end
+end