+++ lib/sourcescrub/utils/veriables.rb
@@ -1,0 +1,17 @@
+# frozen_string_literal: true
+
+module Sourcescrub
+  # Utils
+  module Utils
+    # Veriables
+    module Veriables
+      # Description of operators
+      PEOPLE_ROLES = {
+        'founder' => 'Founder',
+        'employee' => 'Employee',
+        'board_member' => 'Board Member',
+        'advisor' => 'Advisor'
+      }.freeze
+    end
+  end
+end